+++ integration/js/AppQuitVideoTest.js
@@ -3,7 +3,7 @@ const {UserJoinedMeetingCheck, RemoteVideoCheck, UserAuthenticationCheck, Roster
 const {TestUtils} = require('./node_modules/kite-common');
 const SdkBaseTest = require('./utils/SdkBaseTest');
 const {Window} = require('./utils/Window');
-const uuidv4 = require('uuid/v4');
+const { v4: uuidv4 } = require('uuid');
 
 class AppQuitVideoTest extends SdkBaseTest {
   constructor(name, kiteConfig) {